Rename React source directories.
This rename is part of #566 and is a follow-up to #566.
