+++ packages/thumbprint-react/components/Calendar/index.module.scss
