+++ packages/thumbprint-react/components/Calendar/utilities.ts
