+++ packages/thumbprint-react/components/Dropdown/index.module.scss
