+++ packages/thumbprint-react/components/Modal/components/sticky-footer.module.scss
