+++ packages/thumbprint-react/components/Modal/components/transition.jsx
