+++ packages/thumbprint-react/components/Modal/index.jsx
