+++ packages/thumbprint-react/components/Modal/index.module.scss
