+++ packages/thumbprint-react/components/TextArea/index.module.scss
