+++ packages/thumbprint-react/components/TextInput/index.module.scss
