+++ www/src/pages/components/calendar/react/index.mdx
@@ -184,7 +184,7 @@ export const pageQuery = graphql`
         }
         # Get component props by path to component file.
         reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/DatePicker/index.tsx" }
+            relativePath: { eq: "thumbprint-react/components/Calendar/index.tsx" }
         ) {
             ...ReactComponentPropsFragment
         }
