+++ www/src/pages/components/dropdown/react/index.mdx
@@ -174,7 +174,7 @@ export const pageQuery = graphql`
         }
         # Get component props by path to component file.
         reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Select/index.tsx" }
+            relativePath: { eq: "thumbprint-react/components/Dropdown/index.tsx" }
         ) {
             ...ReactComponentPropsFragment
         }
