+++ www/src/pages/components/modal/react/index.mdx
@@ -671,7 +671,7 @@ export const pageQuery = graphql`
         }
         # Get component props by path to component file.
         reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/ModalDefault/index.jsx" }
+            relativePath: { eq: "thumbprint-react/components/Modal/index.jsx" }
         ) {
             ...ReactComponentPropsFragment
         }
