+++ www/src/pages/components/text-area/react/index.mdx
@@ -95,7 +95,7 @@ export const pageQuery = graphql`
         }
         # Get component props by path to component file.
         reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Textarea/index.tsx" }
+            relativePath: { eq: "thumbprint-react/components/TextArea/index.tsx" }
         ) {
             ...ReactComponentPropsFragment
         }
