+++ www/src/pages/components/text-input/react/index.mdx
@@ -151,7 +151,7 @@ export const pageQuery = graphql`
         }
         # Get component props by path to component file.
         reactComponentProps: file(
-            relativePath: { eq: "thumbprint-react/components/Input/index.tsx" }
+            relativePath: { eq: "thumbprint-react/components/TextInput/index.tsx" }
         ) {
             ...ReactComponentPropsFragment
         }
